+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Added emoji count check to profile data validator.
 - Check mention and link counts when creating post.
 - Disable transaction monitor tasks if blockchain integration is disabled.
+- Allow multiple configurations in `blockchains` array.
 
 ## [1.20.0] - 2023-03-07